train.py contains sample Ray code for FSDP but running into local access error
